requirements.txt
opencv-python
numpy
requests
python-dotenv
playwright
rapidfuzz
SpeechRecognition
PyAudio
streamlit